added a benchmark
This patched original version is faster than the original,
and also allows us to store more items, the trade off,
is that you need to set the (aproximated) size of the
keys (That fits on our use case)

Anyway if we have concerns about, the way we are cleaning,
the dict.. (To not block too much time), we can add somehting
like this:

https://github.com/antirez/redis/blob/unstable/src/dict.c#L240

That limits the cleanup for a number of milliseconds, and
also we can start doing this within a threshold. Anyway, after
running the benchamrks, doesn't feel necessary for me.